requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
Django==5.2.7
django-ninja==1.4.5
djangorestframework==3.16.1
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
